+++ include/nat.h
@@ -106,7 +106,7 @@ struct natstr {
     short nat_btu;		/* bureaucratic time units */
     short nat_access;		/* The tick when BTUs were last updated */
     int nat_reserve;		/* military reserves */
-    long nat_money;		/* moola */
+    int nat_money;		/* moola */
     time_t nat_last_login;	/* time of last login, 0 menas never */
     time_t nat_last_logout;	/* time of last logout, 0 means never */
     time_t nat_newstim;		/* date news last read */
